+++ src/functions.php
@@ -8,4 +8,5 @@
 
 require_once __DIR__ . '/make.php';
 require_once __DIR__ . '/Attribute/Requires.php';
+require_once __DIR__ . '/Attribute/Target.php';
 require_once __DIR__ . '/listener.php';